requirements.txt
# Pinned dependencies for reproducing the MLR 555 Group 13 research project.
# Tested with Python 3.10+ on Linux/macOS and Google Colab (Ubuntu 22.04).
#
# IMPORTANT: numpy<2 is required because gym (used transitively by TD-MPC2's
# replay buffer code path) does not yet support numpy 2.0+ ABI.

# Core numerics — pin numpy below 2.0
numpy==1.26.4
scipy>=1.11
pandas>=2.0

# Deep learning
torch>=2.1
tensordict>=0.3
torchrl>=0.3

# Reinforcement learning
gymnasium>=0.29
highway-env>=1.8
stable-baselines3[extra]>=2.2

# TD-MPC2 dependencies (its codebase is cloned separately into third_party/)
hydra-core>=1.3
omegaconf>=2.3
termcolor>=2.3

# Experiment tracking
wandb>=0.16

# Plotting & analysis
matplotlib>=3.7
seaborn>=0.13

# Video / image (optional, for demo recording)
imageio>=2.31
imageio-ffmpeg>=0.4
moviepy>=1.0
Pillow>=10.0

# Notebook support
ipykernel
jupyter
